refactor(incidents): drop migrations, random base36 ids, split state module
- Remove the schema-migration framework. The incident feature is not yet
  released, so there is no persisted data to migrate; the incidents table is
  created lazily on first write.
- Incident ids are now a random u32 used directly as the redb key and shown as
  dash-grouped base36 (e.g. "x7a8-fw2i"). create_incident allocates a fresh,
  unused id (retrying on the unlikely collision); encode/decode helpers live in
  grey_api.
- Split agent/src/state.rs into a state/ module: IncidentStore and ProbeStore
  traits with their impls over State live in their own sub-modules (incidents.rs,
  probes.rs), keeping the gossip/cluster store in mod.rs.

Tests: grey 111 / grey-api 26 / grey-ui 17.
